packages/bookworm/pyproject.toml
[project]
name = "bookworm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.4.0",
    "gunicorn>=23.0.0",
    "whitenoise>=6.7.0",
    "drf-yasg>=1.21.7",
    "django>=5.1.1",
    "pagekeeper",
    "bookcourier",
    "shortuuid>=1.0.13",
    "environs[django]>=11.0.0",
    "psycopg2-binary>=2.9.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
dev-dependencies = []

[tool.uv.sources]
pagekeeper = { workspace = true }
bookcourier = { workspace = true }
